requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.9.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.10
DAWG-Python==0.7.2
elastic-transport==8.15.1
elasticsearch==8.16.0
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.11
numpy==2.0.2
packaging==24.2
preshed==3.0.9
priority==2.0.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
PyMuPDF==1.24.14
pyspellchecker==0.8.1
python-dotenv==1.0.1
Quart==0.19.9
quart-cors==0.7.0
redis==5.2.0
requests==2.32.3
rich==13.9.4
ru_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.8.0/ru_core_news_sm-3.8.0-py3-none-any.whl#sha256=69978d47b43e2c4f329bebdb155e8e9d3861bba1a58ba25551419dae7d7e07fc
setuptools==75.6.0
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.3.2
tqdm==4.67.1
typer==0.15.0
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
